% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyphone.R
\name{sky_get}
\alias{sky_get}
\title{Fetch GitHub Skyline Data}
\usage{
sky_get(user, year = 2020)
}
\arguments{
\item{user}{Character string. A GitHub user whose contributions you want to
fetch.}

\item{year}{Single numeric value. The year for which you want to collect the
contribution count.}
}
\value{
A data.frame/tibble with columns user, year, week, day, date and
count. Row count will be 365 or 366 if a leap year.
}
\description{
Construct a GitHub Skyline API path for a given user and year, fetch and
augment the data, and convert to a tidy tabular format.
}
\examples{
\dontrun{ sky_get(user = "matt-dray", year = 2020) }
}
