% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyphone.R
\name{sky_plot}
\alias{sky_plot}
\title{Plot A GitHub Skyline Dataset}
\usage{
sky_plot(data)
}
\arguments{
\item{data}{A data.frame/tibble. An object output from \code{\link{sky_get}}
containing the GitHub contributions data for a given user in a given
year.}
}
\value{
A gg/ggplot-class object.
}
\description{
Produce a simple, opinionated plot of GitHub Skyline data showing
contribution count over time for a given year.
}
\examples{
\dontrun{
    d <- sky_get(user = "matt-dray", year = 2020)
    sky_plot(d)
    }
}
