% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyphone.R
\name{sky_sonify}
\alias{sky_sonify}
\title{Sonify A GitHub Skyline Dataset}
\usage{
sky_sonify(data, play = TRUE, out_dir = NULL)
}
\arguments{
\item{data}{A data.frame/tibble. An object output from \code{\link{sky_get}}
containing the GitHub contributions data for a given user in a given
year.}

\item{play}{Logical. Should the audio be played? Defaults to TRUE}

\item{out_dir}{Character. The folder to which the audio file will be saved.
Defaults to NULL (no file is saved).}
}
\value{
A WaveMC-class object, plus optional audio output and a .wav file.
}
\description{
Present GitHub Skyline contributions data in audio format using the {sonify}
package.
}
\examples{
\dontrun{
    d <- sky_get(user = "matt-dray", year = 2020)
    sky_sonify (d, play = FALSE, out_dir = "~/Desktop")
    }
}
