% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyphone-package.R
\docType{package}
\name{skyphone-package}
\alias{skyphone}
\alias{skyphone-package}
\title{skyphone: Fetch, Sonify And Plot GitHub Skyline Data}
\description{
Fetch contributions data for a given GitHub user in a given year
    via the GitHub Skyline API, with functions for presenting in audio or
    visual formats.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
